+++ loader/loader.go
@@ -517,14 +517,14 @@ func (l *Loader) Init(ctx context.Context) (err error) {
 
 // Process implements Unit.Process.
 func (l *Loader) Process(ctx context.Context, pr chan pb.ProcessResult) {
-	AddLoaderExitWithErrorCounter(float64(0), l.cfg.Name, l.cfg.SourceID)
+	AddLoaderExitWithErrorCounter(0, l.cfg.Name, l.cfg.SourceID)
 
 	newCtx, cancel := context.WithCancel(ctx)
 	defer cancel()
 
 	l.newFileJobQueue()
 	if err := l.getMydumpMetadata(); err != nil {
-		AddLoaderExitWithErrorCounter(float64(1), l.cfg.Name, l.cfg.SourceID)
+		AddLoaderExitWithErrorCounter(1, l.cfg.Name, l.cfg.SourceID)
 		pr <- pb.ProcessResult{
 			Errors: []*pb.ProcessError{unit.NewProcessError(err)},
 		}
@@ -559,7 +559,7 @@ func (l *Loader) Process(ctx context.Context, pr chan pb.ProcessResult) {
 		if utils.IsContextCanceledError(err) {
 			l.logger.Info("filter out error caused by user cancel")
 		} else {
-			AddLoaderExitWithErrorCounter(float64(1), l.cfg.Name, l.cfg.SourceID)
+			AddLoaderExitWithErrorCounter(1, l.cfg.Name, l.cfg.SourceID)
 			errs = append(errs, unit.NewProcessError(err))
 		}
 	}
